+++ exporter/opentelemetry-exporter-jaeger/src/opentelemetry/exporter/jaeger/__init__.py
@@ -44,6 +44,7 @@ Usage
         # collector_endpoint='http://localhost:14268/api/traces?format=jaeger.thrift',
         # username=xxxx, # optional
         # password=xxxx, # optional
+        # insecure=True, # optional
         # credentials=xxx # optional channel creds
         # transport_format='protobuf' # optional
         # max_tag_value_length=None # optional
@@ -57,6 +58,16 @@ Usage
 
     with tracer.start_as_current_span('foo'):
         print('Hello world!')
+
+You can configure the exporter with the following environment variables:
+
+- :envvar:`OTEL_EXPORTER_JAEGER_USER`
+- :envvar:`OTEL_EXPORTER_JAEGER_PASSWORD`
+- :envvar:`OTEL_EXPORTER_JAEGER_INSECURE`
+- :envvar:`OTEL_EXPORTER_JAEGER_ENDPOINT`
+- :envvar:`OTEL_EXPORTER_JAEGER_CERTIFICATE`
+- :envvar:`OTEL_EXPORTER_JAEGER_AGENT_PORT`
+- :envvar:`OTEL_EXPORTER_JAEGER_AGENT_HOST`
 
 API
 ---
@@ -118,8 +129,8 @@ class JaegerExporter(SpanExporter):
             required.
         password: The password of the Basic Auth if authentication is
             required.
-        credentials: Optional Credentials for server authentication. Set to None to use
-            insecure channel.
+        insecure: True if collector has no encryption or authentication
+        credentials: Credentials for server authentication.
         transport_format: Transport format for exporting spans to collector.
         max_tag_value_length: Max length string attribute values can have. Set to None to disable.
         udp_split_oversized_batches: Re-emit oversized batches in smaller chunks.
@@ -133,6 +144,7 @@ class JaegerExporter(SpanExporter):
         collector_endpoint: Optional[str] = None,
         username: Optional[str] = None,
         password: Optional[str] = None,
+        insecure: Optional[bool] = None,
         credentials: Optional[ChannelCredentials] = None,
         transport_format: Optional[str] = None,
         max_tag_value_length: Optional[int] = None,
@@ -185,6 +197,7 @@ class JaegerExporter(SpanExporter):
         )
         self._collector = None
         self._grpc_client = None
+        self.insecure = insecure
         self.credentials = util._get_credentials(credentials)
         self.transport_format = (
             transport_format.lower()
@@ -200,7 +213,7 @@ class JaegerExporter(SpanExporter):
         endpoint = self.collector_endpoint or DEFAULT_GRPC_COLLECTOR_ENDPOINT
 
         if self._grpc_client is None:
-            if self.credentials is None:
+            if self.insecure:
                 self._grpc_client = CollectorServiceStub(
                     insecure_channel(endpoint)
                 )