+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/export/__init__.py
@@ -115,8 +115,16 @@ class _FlushRequest:
 class BatchSpanProcessor(SpanProcessor):
     """Batch span processor implementation.
 
-    BatchSpanProcessor is an implementation of `SpanProcessor` that
+    `BatchSpanProcessor` is an implementation of `SpanProcessor` that
     batches ended spans and pushes them to the configured `SpanExporter`.
+
+    `BatchSpanProcessor` is configurable with the following environment
+    variables which correspond to constructor parameters:
+
+    - :envvar:`OTEL_BSP_SCHEDULE_DELAY`
+    - :envvar:`OTEL_BSP_MAX_QUEUE_SIZE`
+    - :envvar:`OTEL_BSP_MAX_EXPORT_BATCH_SIZE`
+    - :envvar:`OTEL_BSP_EXPORT_TIMEOUT`
     """
 
     def __init__(