+++ CHANGELOG.md
@@ -52,11 +52,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1684](https://github.com/open-telemetry/opentelemetry-python/pull/1684))
 - Cleanup OTLP exporter compression options, add tests
   ([#1671](https://github.com/open-telemetry/opentelemetry-python/pull/1671))
+- Initial documentation for environment variables
+  ([#1680](https://github.com/open-telemetry/opentelemetry-python/pull/1680))
 
 ### Removed
 - Removed unused `get_hexadecimal_trace_id` and `get_hexadecimal_span_id` methods.
   ([#1675](https://github.com/open-telemetry/opentelemetry-python/pull/1675))
-- Remove insecure param from exporters constructor and `OTEL_EXPORTER_*_ INSECURE` env var
+- Remove `OTEL_EXPORTER_*_ INSECURE` env var
   ([#1682](https://github.com/open-telemetry/opentelemetry-python/pull/1682))
 
 ## [0.18b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v0.18b0) - 2021-02-16